% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Export a Shiny app to a directory}
\usage{
export(appdir, destdir, ..., subdir = "", verbose = FALSE)
}
\arguments{
\item{appdir}{Directory containing the application.}

\item{destdir}{Destination directory.}

\item{...}{Ignored}

\item{subdir}{Subdirectory of \code{destdir} to write the app to.}

\item{verbose}{Print verbose output.}
}
\description{
This function exports a Shiny app to a directory, which can then be served
using \code{httpuv::runStaticServer()}.
}
\examples{
\dontrun{
app_dir <- system.file("examples", "01_hello", package="shiny")
out_dir <- tempfile("shinylive-export")

# Export the app to a directory
export(app_dir, out_dir)
#> Run the following in an R session to serve the app:
#>   httpuv::runStaticServer(<OUT_DIR>, port=8008)

# Serve the exported directory
httpuv::runStaticServer(out_dir, port=8008)
}
}
