% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_ext.R
\name{quarto_ext}
\alias{quarto_ext}
\title{Quarto extension for shinylive}
\usage{
quarto_ext(args = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{args}{Command line arguments passed by the extension. The first
argument can be one of:
\itemize{
\item \code{codeblock-to-json-path}: Prints the path to the \code{codeblock-to-json.js} script.
\item \code{base-deps}: Prints the base dependencies as a JSON array.
\item \code{package-deps}: Prints the package dependencies as a JSON array.
Currently, this returns an empty array as \code{webr} is handling the package
dependencies.
}}
}
\description{
Quarto extension for shinylive
}
