% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinylive-package.R
\docType{package}
\name{shinylive-package}
\alias{shinylive}
\alias{shinylive-package}
\title{shinylive: Run 'shiny' Applications in the Browser}
\description{
Exporting 'shiny' applications with 'shinylive' allows you to run them entirely in a web browser, without the need for a separate R server. The traditional way of deploying 'shiny' applications involves in a separate server and client: the server runs R and 'shiny', and clients connect via the web browser. When an application is deployed with 'shinylive', R and 'shiny' run in the web browser (via 'webR'): the browser is effectively both the client and server for the application. This allows for your 'shiny' application exported by 'shinylive' to be hosted by a static web server.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://posit-dev.github.io/r-shinylive/}
  \item \url{https://github.com/posit-dev/r-shinylive}
  \item Report bugs at \url{https://github.com/posit-dev/r-shinylive/issues}
}

}
\author{
\strong{Maintainer}: Barret Schloerke \email{barret@posit.co} (\href{https://orcid.org/0000-0001-9986-114X}{ORCID})

Authors:
\itemize{
  \item Winston Chang \email{winston@posit.co} (\href{https://orcid.org/0000-0002-1576-2126}{ORCID})
  \item George Stagg \email{george.stagg@posit.co}
  \item Garrick Aden-Buie \email{garrick@posit.co} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
